# swagger.yaml
openapi: 3.0.0
info:
  title: API Banco Principal
  description: |
    API REST completa para sistema bancário digital moderno.
    
    Esta API oferece funcionalidades essenciais para operações bancárias, incluindo:
    - Gerenciamento de contas de usuário
    - Cartões de crédito e débito
    - Transações financeiras (depósitos, transferências, compras)
    - Investimentos e rentabilidade
    - Extratos e relatórios
    
    ## Autenticação
    A API utiliza JWT (JSON Web Tokens) para autenticação. Após o login, inclua o token no header:
    ```
    Authorization: Bearer <seu-token-jwt>
    ```
    
    ## Códigos de Status
    - `200` - Sucesso
    - `201` - Criado com sucesso
    - `400` - Erro de validação
    - `401` - Não autorizado
    - `404` - Recurso não encontrado
    - `500` - Erro interno do servidor
  version: 2.0.0
  contact:
    name: Suporte API Banco Principal
    email: suporte@bancoprincipal.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.bancoprincipal.com
    description: Servidor de produção

paths:
  /health:
    get:
      tags:
        - Sistema
      summary: Verificação de saúde da API
      description: Endpoint para verificar se a API está funcionando corretamente
      responses:
        '200':
          description: API funcionando normalmente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
                  timestamp:
                    type: string
                    format: date-time
                  uptime:
                    type: number

  /api/v1/auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar novo usuário
      description: Cria uma nova conta de usuário no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/login:
    post:
      tags:
        - Autenticação
      summary: Login do usuário
      description: Autentica o usuário e retorna um token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/account/profile:
    get:
      tags:
        - Conta
      summary: Dados do perfil do usuário
      description: Retorna informações da conta do usuário logado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dados do perfil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/account/statement:
    get:
      tags:
        - Conta
      summary: Extrato da conta
      description: Retorna o extrato detalhado da conta com filtros opcionais
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          description: Data inicial (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Data final (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: transactionType
          in: query
          description: Tipo de transação
          schema:
            type: string
            enum: [ALL, DEPOSIT, TRANSFER, CREDIT_PURCHASE, BILL_PAYMENT]
      responses:
        '200':
          description: Extrato da conta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStatement'

  /api/v1/cards:
    get:
      tags:
        - Cartões
      summary: Lista cartões do usuário
      description: Retorna todos os cartões associados ao usuário
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de cartões
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsResponse'

  /api/v1/cards/request:
    post:
      tags:
        - Cartões
      summary: Solicitar cartão de crédito
      description: Solicita um novo cartão de crédito para o usuário
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
      responses:
        '201':
          description: Cartão solicitado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardCreatedResponse'

  /api/v1/transactions/deposit:
    post:
      tags:
        - Transações
      summary: Depósito em conta
      description: Realiza um depósito na conta do usuário
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '201':
          description: Depósito realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/v1/transactions/transfer:
    post:
      tags:
        - Transações
      summary: Transferência entre contas
      description: Realiza transferência para outra conta
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transferência realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/v1/transactions/credit-purchase:
    post:
      tags:
        - Transações
      summary: Compra no cartão de crédito
      description: Realiza uma compra utilizando o cartão de crédito
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPurchaseRequest'
      responses:
        '201':
          description: Compra realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/v1/transactions/pay-bill:
    post:
      tags:
        - Transações
      summary: Pagamento de faturas
      description: Realiza o pagamento de faturas (cartão de crédito, financiamentos, etc.)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillPaymentRequest'
      responses:
        '201':
          description: Pagamento realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/v1/investments/summary:
    get:
      tags:
        - Investimentos
      summary: Resumo de investimentos
      description: Retorna o resumo dos investimentos e rentabilidade do usuário
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Resumo de investimentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentSummary'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegistration:
      type: object
      required:
        - nomeCompleto
        - cpf
        - email
        - senha
        - telefone
      properties:
        nomeCompleto:
          type: string
          example: "João Silva Santos"
        cpf:
          type: string
          pattern: '^[0-9]{11}$'
          example: "12345678901"
        email:
          type: string
          format: email
          example: "joao.silva@email.com"
        senha:
          type: string
          minLength: 8
          example: "MinhaSenh@123"
        telefone:
          type: string
          example: "11987654321"
        endereco:
          type: object
          properties:
            rua:
              type: string
              example: "Rua das Flores, 123"
            cidade:
              type: string
              example: "São Paulo"
            estado:
              type: string
              example: "SP"
            cep:
              type: string
              example: "01234-567"

    LoginRequest:
      type: object
      required:
        - cpf
        - senha
      properties:
        cpf:
          type: string
          example: "12345678901"
        senha:
          type: string
          example: "MinhaSenh@123"

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        usuario:
          $ref: '#/components/schemas/UserProfile'

    UserCreatedResponse:
      type: object
      properties:
        message:
          type: string
          example: "Usuário criado com sucesso"
        usuario:
          type: object
          properties:
            id:
              type: string
              example: "uuid-123"
            nomeCompleto:
              type: string
              example: "João Silva Santos"
            agencia:
              type: string
              example: "0001"
            numeroConta:
              type: string
              example: "123456"

    UserProfile:
      type: object
      properties:
        id:
          type: string
        nomeCompleto:
          type: string
        cpf:
          type: string
        email:
          type: string
        agencia:
          type: string
        numeroConta:
          type: string
        saldo:
          type: number
          format: float
        limiteDebitoDiario:
          type: number
          format: float
        limiteCreditoDiario:
          type: number
          format: float

    AccountStatement:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/UserProfile'
        statement:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        summary:
          type: object
          properties:
            totalEntradas:
              type: number
              format: float
            totalSaidas:
              type: number
              format: float
            saldoInicial:
              type: number
              format: float
            saldoFinal:
              type: number
              format: float

    Transaction:
      type: object
      properties:
        id:
          type: string
        tipo:
          type: string
          enum: [DEPOSIT, TRANSFER, CREDIT_PURCHASE, BILL_PAYMENT]
        valor:
          type: number
          format: float
        descricao:
          type: string
        dataTransacao:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]

    CardsResponse:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'

    Card:
      type: object
      properties:
        id:
          type: string
        numeroCartao:
          type: string
          example: "****-****-****-1234"
        tipoCartao:
          type: string
          enum: [DEBITO, CREDITO]
        limiteCredito:
          type: number
          format: float
        creditoUtilizado:
          type: number
          format: float
        status:
          type: string
          enum: [ATIVO, BLOQUEADO, CANCELADO]

    CardRequest:
      type: object
      required:
        - tipoCartao
        - limiteCredito
      properties:
        tipoCartao:
          type: string
          enum: [CREDITO]
        limiteCredito:
          type: number
          format: float
          example: 5000.00

    CardCreatedResponse:
      type: object
      properties:
        message:
          type: string
          example: "Cartão solicitado com sucesso"
        cartao:
          $ref: '#/components/schemas/Card'

    DepositRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 500.00
        description:
          type: string
          example: "Depósito via transferência"

    TransferRequest:
      type: object
      required:
        - amount
        - destinationAccount
        - destinationAgency
        - destinationName
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 200.00
        destinationAccount:
          type: string
          example: "654321"
        destinationAgency:
          type: string
          example: "0002"
        destinationName:
          type: string
          example: "Maria Santos"
        description:
          type: string
          example: "Transferência PIX"

    CreditPurchaseRequest:
      type: object
      required:
        - amount
        - merchant
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 150.00
        merchant:
          type: string
          example: "Supermercado ABC"
        description:
          type: string
          example: "Compra de produtos"

    BillPaymentRequest:
      type: object
      required:
        - amount
        - billType
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 300.00
        billType:
          type: string
          enum: [CREDIT_CARD, FINANCING, UTILITIES, INSURANCE]
          example: "CREDIT_CARD"
        description:
          type: string
          example: "Pagamento de fatura do cartão"

    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: "txn-uuid-123"
        type:
          type: string
          enum: [DEPOSIT, TRANSFER, CREDIT_PURCHASE, BILL_PAYMENT]
        amount:
          type: number
          format: float
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
        timestamp:
          type: string
          format: date-time
        description:
          type: string

    InvestmentSummary:
      type: object
      properties:
        totalInvestido:
          type: number
          format: float
          example: 10000.00
        totalAtual:
          type: number
          format: float
          example: 10500.00
        rentabilidadeTotal:
          type: number
          format: float
          example: 5.0
        investments:
          type: array
          items:
            $ref: '#/components/schemas/Investment'

    Investment:
      type: object
      properties:
        id:
          type: string
        tipoInvestimento:
          type: string
          enum: [CDB, POUPANCA, TESOURO_DIRETO, FUNDOS]
        valorInvestido:
          type: number
          format: float
        valorAtual:
          type: number
          format: float
        rentabilidade:
          type: number
          format: float
        dataCriacao:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Mensagem de erro"
        message:
          type: string
          example: "Descrição detalhada do erro"
        statusCode:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time

tags:
  - name: Sistema
    description: Endpoints de sistema e monitoramento
  - name: Autenticação
    description: Endpoints de autenticação e autorização
  - name: Conta
    description: Gerenciamento de conta do usuário
  - name: Cartões
    description: Gerenciamento de cartões de crédito e débito
  - name: Transações
    description: Operações financeiras (depósitos, transferências, compras)
  - name: Investimentos
    description: Consulta e gerenciamento de investimentos